StreamExecutionEnvironment and MiniCluster created once
