+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/runner/FlinkTestMain.scala
@@ -1,16 +1,18 @@
 package pl.touk.nussknacker.engine.process.runner
 
 import io.circe.Json
-import org.apache.flink.configuration.Configuration
 import org.apache.flink.runtime.jobgraph.SavepointRestoreSettings
+import org.apache.flink.runtime.minicluster.MiniCluster
+import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.api.test.ScenarioTestData
-import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion, StreamMetaData}
+import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.{AdditionalModelConfigs, DeploymentData}
 import pl.touk.nussknacker.engine.process.compiler.TestFlinkProcessCompilerDataFactory
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
+import pl.touk.nussknacker.engine.process.testmechanism.FlinkStubbedRunner
 import pl.touk.nussknacker.engine.process.{ExecutionConfigPreparer, FlinkJobConfig}
 import pl.touk.nussknacker.engine.testmode.TestProcess.TestResults
 import pl.touk.nussknacker.engine.testmode.{
@@ -18,57 +20,54 @@ import pl.touk.nussknacker.engine.testmode.{
   ResultsCollectingListenerHolder,
   TestServiceInvocationCollector
 }
-import pl.touk.nussknacker.engine.util.MetaDataExtractor
 
 object FlinkTestMain extends FlinkRunner {
 
   def run(
+      miniCluster: MiniCluster,
+      env: StreamExecutionEnvironment,
       modelData: ModelData,
       process: CanonicalProcess,
-      scenarioTestData: ScenarioTestData,
-      configuration: Configuration,
+      scenarioTestData: ScenarioTestData
   ): TestResults[Json] = {
     val processVersion = ProcessVersion.empty.copy(processName =
       ProcessName("snapshot version")
     ) // testing process may be unreleased, so it has no version
     new FlinkTestMain(
+      miniCluster,
+      env,
       modelData,
       process,
       scenarioTestData,
       processVersion,
       DeploymentData.empty.copy(additionalModelConfigs =
         AdditionalModelConfigs(modelData.additionalConfigsFromProvider)
-      ),
-      configuration
+      )
     ).runTest
   }
 
 }
 
 class FlinkTestMain(
+    miniCluster: MiniCluster,
+    env: StreamExecutionEnvironment,
     modelData: ModelData,
     process: CanonicalProcess,
     scenarioTestData: ScenarioTestData,
     processVersion: ProcessVersion,
-    deploymentData: DeploymentData,
-    configuration: Configuration
+    deploymentData: DeploymentData
 ) {
 
-  private val stubbedRunner = new FlinkStubbedRunner(modelData.modelClassLoader, configuration)
+  private val stubbedRunner = new FlinkStubbedRunner(miniCluster, env)
 
   def runTest: TestResults[Json] = {
     val collectingListener = ResultsCollectingListenerHolder.registerTestEngineListener
     try {
       val resultCollector = new TestServiceInvocationCollector(collectingListener)
       val registrar       = prepareRegistrar(collectingListener, scenarioTestData)
-      val parallelism = MetaDataExtractor
-        .extractTypeSpecificDataOrDefault[StreamMetaData](process.metaData, StreamMetaData())
-        .parallelism
-        .getOrElse(1)
-      val env = stubbedRunner.createEnv(parallelism)
 
       registrar.register(env, process, processVersion, deploymentData, resultCollector)
-      stubbedRunner.execute(env, parallelism, process.name, SavepointRestoreSettings.none())
+      stubbedRunner.execute(process.name, SavepointRestoreSettings.none(), modelData.modelClassLoader)
       collectingListener.results
     } finally {
       collectingListener.clean()
