+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/runner/FlinkVerificationMain.scala
@@ -1,55 +1,64 @@
 package pl.touk.nussknacker.engine.process.runner
 
-import org.apache.flink.configuration.Configuration
 import org.apache.flink.runtime.jobgraph.SavepointRestoreSettings
+import org.apache.flink.runtime.minicluster.MiniCluster
+import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
 import pl.touk.nussknacker.engine.ModelData
-import pl.touk.nussknacker.engine.api.{ProcessVersion, StreamMetaData}
+import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.DeploymentData
 import pl.touk.nussknacker.engine.process.compiler.VerificationFlinkProcessCompilerDataFactory
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
+import pl.touk.nussknacker.engine.process.testmechanism.FlinkStubbedRunner
 import pl.touk.nussknacker.engine.process.{ExecutionConfigPreparer, FlinkJobConfig}
-import pl.touk.nussknacker.engine.testmode.{ResultsCollectingListenerHolder, TestRunId, TestServiceInvocationCollector}
-import pl.touk.nussknacker.engine.util.MetaDataExtractor
+import pl.touk.nussknacker.engine.testmode.{ResultsCollectingListenerHolder, TestServiceInvocationCollector}
 
 object FlinkVerificationMain extends FlinkRunner {
 
   def run(
+      miniCluster: MiniCluster,
+      env: StreamExecutionEnvironment,
       modelData: ModelData,
       process: CanonicalProcess,
       processVersion: ProcessVersion,
       deploymentData: DeploymentData,
-      savepointPath: String,
-      configuration: Configuration
+      savepointPath: String
   ): Unit =
-    new FlinkVerificationMain(modelData, process, processVersion, deploymentData, savepointPath, configuration)
-      .runTest()
+    new FlinkVerificationMain(
+      miniCluster,
+      env,
+      modelData,
+      process,
+      processVersion,
+      deploymentData,
+      savepointPath
+    ).runTest()
 
 }
 
 class FlinkVerificationMain(
+    miniCluster: MiniCluster,
+    env: StreamExecutionEnvironment,
     modelData: ModelData,
     process: CanonicalProcess,
     processVersion: ProcessVersion,
     deploymentData: DeploymentData,
-    savepointPath: String,
-    configuration: Configuration
+    savepointPath: String
 ) {
 
-  private val stubbedRunner = new FlinkStubbedRunner(modelData.modelClassLoader, configuration)
+  private val stubbedRunner = new FlinkStubbedRunner(miniCluster, env)
 
   def runTest(): Unit = {
     val collectingListener = ResultsCollectingListenerHolder.registerTestEngineListener
     val resultCollector    = new TestServiceInvocationCollector(collectingListener)
     val registrar          = prepareRegistrar()
-    val parallelism = MetaDataExtractor
-      .extractTypeSpecificDataOrDefault[StreamMetaData](process.metaData, StreamMetaData())
-      .parallelism
-      .getOrElse(1)
-    val env = stubbedRunner.createEnv(parallelism)
 
     registrar.register(env, process, processVersion, deploymentData, resultCollector)
-    stubbedRunner.execute(env, parallelism, process.name, SavepointRestoreSettings.forPath(savepointPath, true))
+    stubbedRunner.execute(
+      process.name,
+      SavepointRestoreSettings.forPath(savepointPath, true),
+      modelData.modelClassLoader
+    )
   }
 
   protected def prepareRegistrar(): FlinkProcessRegistrar = {
