+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/testsmechanism/FlinkProcessVerifier.scala
@@ -1,18 +1,21 @@
-package pl.touk.nussknacker.engine.management
+package pl.touk.nussknacker.engine.management.testsmechanism
 
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.flink.configuration.Configuration
 import pl.touk.nussknacker.engine.ModelData
-import pl.touk.nussknacker.engine.api.ProcessVersion
+import pl.touk.nussknacker.engine.api.{ProcessVersion, StreamMetaData}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.deployment.DeploymentData
-import pl.touk.nussknacker.engine.util.ReflectiveMethodInvoker
+import pl.touk.nussknacker.engine.util.{MetaDataExtractor, ReflectiveMethodInvoker}
 
 import scala.concurrent.Future
+import scala.util.Using
 import scala.util.control.NonFatal
 
 class FlinkProcessVerifier(modelData: ModelData) extends LazyLogging {
 
+  // We use reflection to avoid bundling of flinkExecutor.jar inside flinkDeploymentManager assembly jar
+  // TODO: use provided dependency instead
   private val methodInvoker = new ReflectiveMethodInvoker[Unit](
     modelData.modelClassLoader.classLoader,
     "pl.touk.nussknacker.engine.process.runner.FlinkVerificationMain",
@@ -24,17 +27,30 @@ class FlinkProcessVerifier(modelData: ModelData) extends LazyLogging {
       canonicalProcess: CanonicalProcess,
       savepointPath: String
   ): Future[Unit] = {
+    val parallelism = MetaDataExtractor
+      .extractTypeSpecificDataOrDefault[StreamMetaData](canonicalProcess.metaData, StreamMetaData())
+      .parallelism
+      .getOrElse(1)
     val processId = processVersion.processName
     try {
       logger.info(s"Starting to verify $processId")
-      methodInvoker.invokeStaticMethod(
-        modelData,
-        canonicalProcess,
-        processVersion,
-        DeploymentData.empty,
-        savepointPath,
-        new Configuration()
-      )
+      // TODO: reuse a single mini cluster between each verifications
+      Using.resource(TestsMechanismMiniClusterFactory.createConfiguredMiniCluster(parallelism)) { miniCluster =>
+        val env = TestsMechanismStreamExecutionEnvironmentFactory.createStreamExecutionEnvironment(
+          parallelism,
+          new Configuration()
+        )
+
+        methodInvoker.invokeStaticMethod(
+          miniCluster,
+          env,
+          modelData,
+          canonicalProcess,
+          processVersion,
+          DeploymentData.empty,
+          savepointPath
+        )
+      }
       logger.info(s"Verification of $processId successful")
       Future.successful(())
     } catch {
