+++ engine/flink/management/src/test/resources/META-INF/services/pl.touk.nussknacker.engine.api.process.ProcessConfigCreator
@@ -1,0 +1,1 @@
+pl.touk.nussknacker.engine.management.testsmechanism.SimpleProcessConfigCreator
