+++ engine/flink/management/src/test/scala/pl/touk/nussknacker/engine/management/testsmechanism/FlinkProcessTestRunnerSpec.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.engine.process.runner
+package pl.touk.nussknacker.engine.management.testsmechanism
 
 import com.typesafe.config.{Config, ConfigFactory, ConfigValueFactory}
 import io.circe.Json
@@ -7,50 +7,38 @@ import org.apache.flink.runtime.client.JobExecutionException
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.{BeforeAndAfterEach, Inside, OptionValues}
-import pl.touk.nussknacker.engine.api.component.{
-  ComponentAdditionalConfig,
-  DesignerWideComponentId,
-  ParameterAdditionalUIConfig
-}
-import pl.touk.nussknacker.engine.api.parameter.{
-  ParameterName,
-  ParameterValueCompileTimeValidation,
-  ValueInputWithDictEditor
-}
+import pl.touk.nussknacker.engine.api.component.{ComponentAdditionalConfig, DesignerWideComponentId, ParameterAdditionalUIConfig}
+import pl.touk.nussknacker.engine.api.parameter.{ParameterName, ParameterValueCompileTimeValidation, ValueInputWithDictEditor}
 import pl.touk.nussknacker.engine.api.process.ComponentUseCase
 import pl.touk.nussknacker.engine.api.test.{ScenarioTestData, ScenarioTestJsonRecord}
-import pl.touk.nussknacker.engine.api.{DisplayJsonWithEncoder, FragmentSpecificData, MetaData}
+import pl.touk.nussknacker.engine.api.{DisplayJsonWithEncoder, FragmentSpecificData, MetaData, StreamMetaData}
 import pl.touk.nussknacker.engine.build.{GraphBuilder, ScenarioBuilder}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.canonicalgraph.canonicalnode.FlatNode
 import pl.touk.nussknacker.engine.compile.FragmentResolver
-import pl.touk.nussknacker.engine.flink.test.{
-  FlinkTestConfiguration,
-  RecordingExceptionConsumer,
-  RecordingExceptionConsumerProvider
-}
+import pl.touk.nussknacker.engine.deployment.AdditionalModelConfigs
+import pl.touk.nussknacker.engine.flink.test.{FlinkTestConfiguration, RecordingExceptionConsumer, RecordingExceptionConsumerProvider}
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{FragmentClazzRef, FragmentParameter}
 import pl.touk.nussknacker.engine.graph.node.{Case, FragmentInputDefinition, FragmentOutputDefinition}
+import pl.touk.nussknacker.engine.management.testsmechanism.FlinkProcessTestRunnerSpec.{fragmentWithValidationName, processWithFragmentParameterValidation}
 import pl.touk.nussknacker.engine.process.helpers.SampleNodes._
-import pl.touk.nussknacker.engine.process.runner.FlinkTestMainSpec.{
-  fragmentWithValidationName,
-  processWithFragmentParameterValidation
-}
 import pl.touk.nussknacker.engine.testmode.TestProcess._
-import pl.touk.nussknacker.engine.util.ThreadUtils
-import pl.touk.nussknacker.engine.{ModelConfigs, ModelData}
-import pl.touk.nussknacker.engine.deployment.AdditionalModelConfigs
-import pl.touk.nussknacker.engine.testing.LocalModelData
+import pl.touk.nussknacker.engine.util.{MetaDataExtractor, ThreadUtils}
 import pl.touk.nussknacker.engine.util.loader.ModelClassLoader
+import pl.touk.nussknacker.engine.{ModelConfigs, ModelData}
 
 import java.util.{Date, UUID}
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
 import scala.concurrent.{Await, Future}
-import scala.jdk.CollectionConverters._
 
-class FlinkTestMainSpec extends AnyWordSpec with Matchers with Inside with BeforeAndAfterEach with OptionValues {
+class FlinkProcessTestRunnerSpec
+    extends AnyWordSpec
+    with Matchers
+    with Inside
+    with BeforeAndAfterEach
+    with OptionValues {
 
   import pl.touk.nussknacker.engine.spel.SpelExtension._
   import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
@@ -244,7 +232,6 @@ class FlinkTestMainSpec extends AnyWordSpec with Matchers with Inside with Befor
       val nodeResults = results.nodeResults
 
       nodeResults(sourceNodeId) should have length 5
-
     }
 
     "detect errors" in {
@@ -776,7 +763,13 @@ class FlinkTestMainSpec extends AnyWordSpec with Matchers with Inside with Befor
       resolveConfigs = false
     )
     ThreadUtils.withThisAsContextClassLoader(getClass.getClassLoader) {
-      FlinkTestMain.run(modelData, process, scenarioTestData, FlinkTestConfiguration.setupMemory(new Configuration))
+      // TODO: reuse this instance between all test cases
+      val parallelism = MetaDataExtractor
+        .extractTypeSpecificDataOrDefault[StreamMetaData](process.metaData, StreamMetaData())
+        .parallelism
+        .getOrElse(1)
+      new FlinkProcessTestRunner(modelData, parallelism, FlinkTestConfiguration.setupMemory(new Configuration))
+        .runTests(process, scenarioTestData)
     }
   }
 
@@ -812,7 +805,7 @@ class FlinkTestMainSpec extends AnyWordSpec with Matchers with Inside with Befor
 
 }
 
-object FlinkTestMainSpec {
+object FlinkProcessTestRunnerSpec {
   private val fragmentWithValidationName = "fragmentWithValidation"
 
   private val processWithFragmentParameterValidation: CanonicalProcess = {
